having issues with rerender on url change
