+++ src/components/Articles.jsx
@@ -1,9 +1,9 @@
 import { useEffect, useState } from "react";
 import { getArticles } from "../utils/api";
-import { Link } from "react-router-dom";
-import sortFunc from "../utils/helper-funcs";
-import { useSearchParams } from "react-router-dom";
+import sortFunc from "../utils/newSort";
+import { useSearchParams, Link, useLocation } from "react-router-dom";
 import RouteMissing from "./Route-missing";
+import Loading from "./Loading";
 
 const Articles = () => {
   const [articles, setArticles] = useState([]);
@@ -13,23 +13,31 @@ const Articles = () => {
   const [byVotes, setByVotes] = useState(false);
   const [err, setErr] = useState(null);
   const [searchParams] = useSearchParams();
+
+  let url = useLocation();
+
+  console.log(url);
+
   useEffect(() => {
     let topic = searchParams.get("topic");
     if (topic) setSeeOnly(topic);
     getArticles(sortBy, order, seeOnly)
       .then((data) => {
         if (byVotes) {
-          sortFunc(data.articles);
+          sortFunc(data.articles, order);
           setArticles(data.articles);
         } else setArticles(data.articles);
       })
       .catch((error) => {
         setErr(error);
       });
-  }, [sortBy, order, seeOnly, byVotes]);
+  }, [sortBy, order, seeOnly, byVotes, url]);
   if (err) {
     console.log(err);
     return <RouteMissing />;
+  }
+  if (articles.length === 0) {
+    return <Loading />;
   }
   return (
     <section>
